+++ sticky_mitten_avatar/util.py
@@ -2,15 +2,10 @@ import numpy as np
 from typing import Dict, List, TypeVar, Type, Optional, Tuple
 from tdw.output_data import OutputData, Transforms, Rigidbodies, Bounds, Images, SegmentationColors, Volumes, Raycast, \
     CompositeObjects, CameraMatrices, Environments, Collision, EnvironmentCollision, Overlap, Version
-from pathlib import Path
-from pkg_resources import resource_filename
 
-# The directory of the occupancy_maps map files.
-OCCUPANCY_MAP_DIRECTORY = Path(resource_filename(__name__, "occupancy_maps"))
-# The path to the scene bounds data.
-SCENE_BOUNDS_PATH = OCCUPANCY_MAP_DIRECTORY.joinpath("scene_bounds.json")
-# The path to the spawn positions.
-SPAWN_POSITIONS_PATH = OCCUPANCY_MAP_DIRECTORY.joinpath("spawn_positions.json")
+
+# The size of each occupancy grid cell.
+OCCUPANCY_CELL_SIZE = 0.25
 
 T = TypeVar("T", bound=OutputData)
 # Output data types mapped to their IDs.