+++ util/spawn_mapper.py
@@ -1,10 +1,10 @@
-from pathlib import Path
 import json
 import numpy as np
 from tdw.controller import Controller
 from tdw.output_data import Environments
 from tdw.tdw_utils import TDWUtils
-from sticky_mitten_avatar.util import get_data, SCENE_BOUNDS_PATH, SPAWN_POSITIONS_PATH
+from sticky_mitten_avatar.util import get_data, OCCUPANCY_CELL_SIZE
+from sticky_mitten_avatar.paths import SCENE_BOUNDS_PATH, SPAWN_POSITIONS_PATH, OCCUPANCY_MAP_DIRECTORY
 
 """
 Calculate avatar spawn positions per scene per layout.
@@ -15,7 +15,6 @@ Rooms are determined with Environments output data.
 
 
 if __name__ == "__main__":
-    p = Path("../sticky_mitten_avatar/occupancy_maps")
     sbd = json.loads(SCENE_BOUNDS_PATH.read_text())
 
     c = Controller()
@@ -43,15 +42,15 @@ if __name__ == "__main__":
             spawn_positions[scene][layout] = list()
             for center in centers:
                 # Load the occupancy map.
-                occ = np.load(str(p.joinpath(f"{scene}_{layout}.npy").resolve()))
+                occ = np.load(str(OCCUPANCY_MAP_DIRECTORY.joinpath(f"{scene}_{layout}.npy").resolve()))
 
                 # Get the free position on the map closest to the center of the room.
                 min_distance = 1000
                 min_position = None
                 for ix, iy in np.ndindex(occ.shape):
                     if occ[ix, iy] == 1:
-                        x = scene_bounds["x_min"] + (ix * 0.25)
-                        z = scene_bounds["z_min"] + (iy * 0.25)
+                        x = scene_bounds["x_min"] + (ix * OCCUPANCY_CELL_SIZE)
+                        z = scene_bounds["z_min"] + (iy * OCCUPANCY_CELL_SIZE)
                         pos = np.array([x, 0, z])
                         d = np.linalg.norm(pos - center)
                         if d < min_distance: